+++ src/main/java/rpn/Expression/Operation/PlusOperation.java
@@ -13,7 +13,7 @@ public class PlusOperation extends AbstractOperation {
 
 
     @Override
-    public void operate(String expression, ArrayDeque<Long> result) throws UnsufficientArgumentException {
+    public void operate(String expression, ArrayDeque<Double> result) throws UnsufficientArgumentException {
         if(result.size()<2)
             throw new UnsufficientArgumentException();
         result.push( result.pop() + result.pop() );
